@@ -4,7 +4,7 @@
     {
         public const string Card = "Kort";
         public const string Recurring = "Recurring";
-        public const string Swish = "Swish";
+        public const string Swish = "swish";
         public const string Invoice = "Faktura";
         public const string Trustly = "Bankbetalning";
     }
