@@ -2,8 +2,7 @@
 using NUnit.Framework;
 using Sample.AspNetCore.SystemTests.Test.Helpers;
 using SwedbankPay.Sdk;
-using SwedbankPay.Sdk.PaymentInstruments;
-using SwedbankPay.Sdk.PaymentOrders;
+using SwedbankPay.Sdk.PaymentOrder;
 using System.Linq;
 
 namespace Sample.AspNetCore.SystemTests.Test.PaymentTests.PaymentOrder.Anonymous
@@ -24,28 +23,28 @@ namespace Sample.AspNetCore.SystemTests.Test.PaymentTests.PaymentOrder.Anonymous
             Assert.DoesNotThrowAsync(async () => {
 
                 GoToOrdersPage(products, payexInfo, Checkout.Option.Anonymous)
-                    .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderCancel)].Should.BeVisible()
-                    .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderCapture)].Should.BeVisible()
-                    .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.PaidPaymentOrder)].Should.BeVisible()
+                    .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.Cancel)].Should.BeVisible()
+                    .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.Capture)].Should.BeVisible()
+                    // .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.PaidPaymentOrder)].Should.BeVisible()
                     .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Clear.ClickAndGo();
 
                 var order = await SwedbankPayClient.PaymentOrders.Get(_link, PaymentOrderExpand.All);
 
                 // Global Order
                 Assert.That(order.PaymentOrder.Amount.InLowestMonetaryUnit, Is.EqualTo(products.Select(x => x.UnitPrice * x.Quantity).Sum()));
-                Assert.That(order.PaymentOrder.Currency.ToString(), Is.EqualTo("SEK"));
-                Assert.That(order.PaymentOrder.State, Is.EqualTo(State.Ready));
+                Assert.That(order.PaymentOrder.Currency, Is.EqualTo("SEK"));
+                Assert.That(order.PaymentOrder.Status, Is.EqualTo(Status.Ready));
 
                 // Operations
                 Assert.That(order.Operations[LinkRelation.CreatePaymentOrderReversal], Is.Null);
-                Assert.That(order.Operations[LinkRelation.CreateCancellation], Is.Not.Null);
+                Assert.That(order.Operations[LinkRelation.CreatePaymentOrderCancel], Is.Not.Null);
                 Assert.That(order.Operations[LinkRelation.CreatePaymentOrderCapture], Is.Not.Null);
-                Assert.That(order.Operations[LinkRelation.PaidPaymentOrder], Is.Not.Null);
+                // Assert.That(order.Operations[LinkRelation.PaidPaymentOrder], Is.Not.Null);
 
                 // Transactions
-                Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.Count, Is.EqualTo(1));
-                Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Authorization).State,
-                            Is.EqualTo(State.Completed));
+                Assert.That(order.PaymentOrder.Paid.Details, Is.Not.Null);
+                // Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Authorization).State,
+                //             Is.EqualTo(State.Completed));
 
                 // Order Items
 
@@ -69,31 +68,31 @@ namespace Sample.AspNetCore.SystemTests.Test.PaymentTests.PaymentOrder.Anonymous
             Assert.DoesNotThrowAsync(async () => {
 
                 GoToOrdersPage(products, payexInfo, Checkout.Option.Anonymous)
-                    .RefreshPageUntil(x => x.Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderCancel)].IsVisible, 60, 10)
-                    .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderCancel)].Should.BeVisible()
-                    .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.CreatePaymentOrderCapture)].Should.BeVisible()
-                    .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.PaidPaymentOrder)].Should.BeVisible()
+                    .RefreshPageUntil(x => x.Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.Cancel)].IsVisible, 60, 10)
+                    .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.Cancel)].Should.BeVisible()
+                    .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.Capture)].Should.BeVisible()
+                    // .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Actions.Rows[y => y.Name.Value.Contains(PaymentOrderResourceOperations.PaidPaymentOrder)].Should.BeVisible()
                     .Orders[y => y.Attributes["data-paymentorderlink"] == _referenceLink].Clear.ClickAndGo();
 
                 var order = await SwedbankPayClient.PaymentOrders.Get(_link, PaymentOrderExpand.All);
 
                 // Global Order
                 Assert.That(order.PaymentOrder.Amount.InLowestMonetaryUnit, Is.EqualTo(products.Select(x => x.UnitPrice * x.Quantity).Sum()));
-                Assert.That(order.PaymentOrder.Currency.ToString(), Is.EqualTo("SEK"));
-                Assert.That(order.PaymentOrder.State, Is.EqualTo(State.Ready));
+                Assert.That(order.PaymentOrder.Currency, Is.EqualTo("SEK"));
+                Assert.That(order.PaymentOrder.Status, Is.EqualTo(Status.Ready));
 
                 // Operations
                 Assert.That(order.Operations[LinkRelation.CreatePaymentOrderReversal], Is.Null);
-                Assert.That(order.Operations[LinkRelation.CreateCancellation], Is.Not.Null);
+                Assert.That(order.Operations[LinkRelation.CreatePaymentOrderCancel], Is.Not.Null);
                 Assert.That(order.Operations[LinkRelation.CreatePaymentOrderCapture], Is.Not.Null);
-                Assert.That(order.Operations[LinkRelation.PaidPaymentOrder], Is.Not.Null);
+                // Assert.That(order.Operations[LinkRelation.PaidPaymentOrder], Is.Not.Null);
 
                 // Transactions
-                Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.Count, Is.EqualTo(2));
-                Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Initialization).State,
-                            Is.EqualTo(State.Completed));
-                Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Authorization).State,
-                            Is.EqualTo(State.Completed));
+                Assert.That(order.PaymentOrder.Paid.Details, Is.Not.Null);
+                // Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Initialization).State,
+                //             Is.EqualTo(State.Completed));
+                // Assert.That(order.PaymentOrder.CurrentPayment.Payment.Transactions.TransactionList.First(x => x.Type == TransactionType.Authorization).State,
+                //             Is.EqualTo(State.Completed));
 
                 // Order Items
                 Assert.That(order.PaymentOrder.OrderItems.OrderItemList.Count, Is.EqualTo(products.Count()));
